Various minor bug fixes for 3.0rc14 release.
	modified:   idas.chox.core/src/main/java/idas/chox/core/model/HireMonitoringDetail.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/bre/rules/LabourCostBusinessRule.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/bre/util/ClaimCalcHelper.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/bre/util/mathHelper.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/reports/viewdata/ClaimFileReportData.java
	modified:   idas.chox.web/src/main/resources/reports/template_InvoiceReport.xls

	modified:   Configuration/DBScript/Phase3Sprint1.6.txt

+++ idas.chox.core/src/main/java/idas/chox/core/model/HireMonitoringDetail.java
@@ -44,7 +44,7 @@ public class HireMonitoringDetail extends Entity implements Serializable {
      * This attribute maps to the column labour_cost in the hire_monitoring_detail table.
      */
     private BigDecimal labourRate;
-    private Integer labourHour;
+    private BigDecimal labourHour;
     private BigDecimal labourCost;
     private String nonProvisionReason;
     protected Date nextReviewDate;
@@ -220,11 +220,11 @@ public class HireMonitoringDetail extends Entity implements Serializable {
         this.labourRate = labourRate;
     }
 
-    public Integer getLabourHour() {
+    public BigDecimal getLabourHour() {
         return labourHour;
     }
 
-    public void setLabourHour(Integer labourHour) {
+    public void setLabourHour(BigDecimal labourHour) {
         this.labourHour = labourHour;
     }
